+++ CHANGE_LOG.md
@@ -5,11 +5,15 @@
 2. Move more event trigger to intermediary.
 3. Add 'attachPreparation' in event context, used to pre-vary raw inputted args.
 4. Add dynamic register APIs for item, block and entity.
+5. Add more minecraft class import.
+6. Clear all event subscribes on reloading data packet
 
 ## New event support
 Add events support:
 1. ITEM_USE
 2. ITEM_USED
+3. RECEIVE_CHUNK
+3. RECEIVED_CHUNK
 
 # Bedrock event
 1. Add 'dimension' field to bedrock entity. (``Entity#dimension``)